+++ src/layouts/auth/index.js
@@ -63,7 +63,7 @@ export default function Auth() {
                 {getRoutes(routes)}
                 <Redirect
                   from='/auth'
-                  to='/auth/sign-in/default
+                  to='/auth/sign-in
                   '
                 />
               </Switch>